# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/chats/group": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new group chat with multiple members and an optional avatar.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Create Group Chat",
                "parameters": [
                    {
                        "description": "Group creation",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateGroupChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ChatListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/chats/group/invite/{inviteCode}": {
            "get": {
                "description": "Get basic group info using an invite code. Useful for previewing before joining.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get Group Preview by Invite Code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite Code",
                        "name": "inviteCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.GroupPreviewDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/chats/group/join/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Join a private or public group using an invite code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Join Group by Invite Code",
                "parameters": [
                    {
                        "description": "Join Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.JoinGroupByInviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ChatListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helper.ResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "model.ChatListResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "Avatar URL of the group or the other user",
                    "type": "string"
                },
                "description": {
                    "description": "Description of the group",
                    "type": "string"
                },
                "hidden_at": {
                    "description": "Timestamp when the current user hid the chat",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invite_code": {
                    "description": "Invite code for the group, available if public or for admins",
                    "type": "string"
                },
                "invite_expires_at": {
                    "description": "Expiration timestamp for the invite code",
                    "type": "string"
                },
                "is_blocked_by_me": {
                    "description": "Indicates if the current user has blocked the other user",
                    "type": "boolean"
                },
                "is_blocked_by_other": {
                    "description": "Indicates if the other user has blocked the current user",
                    "type": "boolean"
                },
                "is_online": {
                    "description": "Indicates if the other user is currently online",
                    "type": "boolean"
                },
                "is_public": {
                    "description": "Indicates if the group is public",
                    "type": "boolean"
                },
                "last_message": {
                    "description": "Preview of the last message in the chat",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MessageResponse"
                        }
                    ]
                },
                "last_read_at": {
                    "description": "Timestamp when the current user last read the chat",
                    "type": "string"
                },
                "member_count": {
                    "description": "Total number of members in the group",
                    "type": "integer"
                },
                "my_role": {
                    "description": "Role of the current user in the group (owner, admin, member)",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the group or the other user in a private chat",
                    "type": "string"
                },
                "other_last_read_at": {
                    "description": "Timestamp when the other user last read the chat",
                    "type": "string"
                },
                "other_user_id": {
                    "description": "ID of the other user in a private chat",
                    "type": "string"
                },
                "other_user_is_banned": {
                    "description": "Indicates if the other user is currently banned",
                    "type": "boolean"
                },
                "other_user_is_deleted": {
                    "description": "Indicates if the other user's account has been deleted",
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "unread_count": {
                    "description": "Number of unread messages for the current user",
                    "type": "integer"
                }
            }
        },
...
        "model.CreateGroupChatRequest": {
            "type": "object",
            "required": [
                "member_ids",
                "name"
            ],
            "properties": {
                "avatar_media_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "is_public": {
                    "type": "boolean"
                },
                "member_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
...
    }
}